rename to `xarray-regridding`
